+++ ConcernsCaseWork/Service.TRAMS/RecordRatingHistory/RecordRatingHistoryDto.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Numerics;
 using System.Text.Json.Serialization;
 
 namespace Service.TRAMS.RecordRatingHistory
@@ -10,13 +9,13 @@ namespace Service.TRAMS.RecordRatingHistory
 		public DateTimeOffset CreatedAt { get; }
 
 		[JsonPropertyName("record_urn")]
-		public BigInteger RecordUrn { get; }
+		public long RecordUrn { get; }
 		
 		[JsonPropertyName("rating_urn")]
-		public BigInteger RatingUrn { get; }
+		public long RatingUrn { get; }
 		
 		[JsonConstructor]
-		public RecordRatingHistoryDto(DateTimeOffset createdAt, BigInteger recordUrn, BigInteger ratingUrn) => 
+		public RecordRatingHistoryDto(DateTimeOffset createdAt, long recordUrn, long ratingUrn) => 
 			(CreatedAt, RecordUrn, RatingUrn) = (createdAt, recordUrn, ratingUrn);
 	}
 }
